project/requirements.txt
asgiref==3.7.2
Django==3.2.21
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.7
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
